pyproject.toml
[tool.poetry]
name = "rockrunde"
version = "0.1.0"
description = ""
authors = ["Markus Nutz <markus.nutz@metro-markets.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
segno = "^1.5.2"
tqdm = "^4.66.1"
spotipy = "^2.23.0"
black = "^23.10.0"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
pillow = "^10.1.0"
xlrd = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
